+++ src/main/java/com/admin_expenses/admin_expenses/application/dto/PurchaseResponseDTO.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -11,7 +13,7 @@ public class PurchaseResponseDTO {
     private Long id;
     private String productName;
     private Integer quantity;
-    private Double installmentAmount;
+    private BigDecimal installmentAmount;
     private String purchaseType;
     private Integer fees;
     private CardResponseDTO card;
